Simplify: merge run_trial/run_trial_with_pitch, fix uint16_t mismatch
- test_dead_reckoning.c: merge run_trial and run_trial_with_pitch into a
  single run_trial(seed, bias, pitch_deg) — the two functions shared ~90
  lines of identical flight simulation logic (motor model, drag, FSM,
  apogee detection); only the body-frame projection differed.
  Pass 0.0f for straight-up flights, non-zero for pitch deviation tests.
  File shrinks from 570 → 372 lines.

- main.c: change run_calibration() loop variable i and counters valid_a,
  valid_b from uint16_t → uint32_t to match DR_CalibrateBias's uint32_t n
  parameter (widened in previous commit); removes the type mismatch and
  the now-redundant explicit (uint32_t) cast at the call site.

+++ src/main.c
@@ -229,8 +229,8 @@ void RocketDR_TimerB_ISR(void)   /* 100 Hz */
 static void run_calibration(void)
 {
     IMU_Data_t d;
-    uint16_t i;
-    uint16_t valid_a = 0u, valid_b = 0u;
+    uint32_t i;
+    uint32_t valid_a = 0u, valid_b = 0u;
 
     /* Collect CALIB_SAMPLE_COUNT samples at 208 Hz (~5 s).
      * Simultaneously accumulate gyro pitch-axis readings for bias estimation:
